pyproject.toml
[tool.poetry]
name = "kalshi_edge"
version = "0.1.0"
description = "Kalshi event-contract backtesting toolkit"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "kalshi_edge" }]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.6"
psycopg2-binary = "^2.9"
httpx = "^0.26"
schedule = "^1.2.0"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
jinja2 = "^3.1.0"
kalshi-python = "^2.1.0"
cryptography = "^46.0.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
ruff = "^0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]
